% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_nlme_pars.R
\name{est_nlme_pars}
\alias{est_nlme_pars}
\title{Estimate parameters in NLME model}
\usage{
est_nlme_pars(
  Sigma,
  info_data,
  id,
  RespLog,
  model1,
  n,
  fixed_names1,
  missing_vars,
  uniqueID,
  S,
  random_effects,
  covariance.model,
  X1i_list,
  X2i_list,
  all_cases,
  par_list,
  mc.cores,
  ptol = 0.01
)
}
\description{
Estimate parameters in NLME model
}
