% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{evalMat}
\alias{evalMat}
\title{Evaluate a matrix given texts}
\usage{
evalMat(mat, data = NULL, par_val = NULL, byrow = TRUE)
}
\arguments{
\item{mat}{a matrix of texts}

\item{data}{data used for evaluation}

\item{par_val}{parameter values needed for evaluation}

\item{byrow}{logical. If TRUE (the default) the matrix is filled by rows.}
}
\description{
Evaluate a matrix given texts
}
