% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log_density.R
\name{get_log_density}
\alias{get_log_density}
\title{Get log density function of (nonlinear) (generalized) linear model or survial model}
\usage{
get_log_density(model_object)
}
\description{
Get log density function of (nonlinear) (generalized) linear model or survial model
}
