% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jm_stem.R
\name{jm_stem}
\alias{jm_stem}
\title{Fit joint model of longitudinal and survival data using Stochastic EM algorithm}
\usage{
jm_stem(
  model1,
  model2,
  covariance.model,
  id,
  continuing = F,
  ini_dat = NULL,
  ini_par_list = NULL,
  sample_Sigma = NULL,
  iterMax = 2500,
  burn_in = 50,
  mcem_par = list(thin = 1, mc_burn = 0, S = 1),
  mc.cores = 1
)
}
\arguments{
\item{model1}{NLME model for longitudinal data}

\item{model2}{Cox PH model for survival data}

\item{covariance.model}{a list specifying variance-covariance matrix structure for random parameters}

\item{id}{a character string specifying the clusters}

\item{continuing}{Is it a continuous run from the previous output of jm_stem()}

\item{ini_dat}{Set imputed longitudinal and survival data from previous output as initial datasets}

\item{ini_par_list}{Set latest parameter estimates from previous output as initial estimates}

\item{sample_Sigma}{variance-covariance matrix for sampling random effects}

\item{iterMax}{number of iterations to run}

\item{burn_in}{burn-in period for MH sampling}

\item{mcem_par}{a list of arguments for MCEM algorithm, including thin for thinning out MC chains,
mc_burn for burn-in period, and S for MC sample size. By default, thin=1, mc_burn=0, and S=1 for StEM algorithm.}

\item{mc.cores}{number of cores used for parallel computing}
}
\description{
Fit joint model of longitudinal and survival data using Stochastic EM algorithm
}
