% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_sampler.R
\name{MC_sampler}
\alias{MC_sampler}
\title{Draw MC samples using Metropolis-Hastings}
\usage{
MC_sampler(
  posterior_theta,
  par_list,
  Sigma,
  sample_Sigma,
  dup_dat1,
  dup_dat2,
  index1,
  index2,
  fixed_names1,
  model1,
  model2,
  random_effects,
  RespLog,
  uniqueID,
  missing_vars,
  id,
  covariance.model,
  X1i_list,
  betahat,
  n,
  loc,
  S = 1,
  thin = 1,
  burn_in = 1,
  mc.cores,
  stopm = 100,
  mc_burn = 0,
  mcpackage = c("MHadaptive", "adaptMCMC", "MCMCpack")
)
}
\description{
Draw MC samples using Metropolis-Hastings
}
