% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metro_Hastings.R
\name{Metro_Hastings}
\alias{Metro_Hastings}
\title{MH sampling funtion from "MHadaptive" package}
\usage{
Metro_Hastings(
  li_func,
  pars,
  prop_sigma = NULL,
  par_names = NULL,
  iterations = 50000,
  burn_in = 1000,
  adapt_par = c(100, 20, 0.5, 0.75),
  quiet = FALSE,
  ...
)
}
\description{
MH sampling funtion from "MHadaptive" package
}
