% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{posterior_theta}
\alias{posterior_theta}
\title{Un-normalized log posterior of random effects theta_i}
\usage{
posterior_theta(
  xx,
  xx_names,
  par_list,
  long_loglike,
  surv_loglike,
  model1,
  sub_dat1,
  sub_dat2,
  Sigma,
  loc,
  theta_mu
)
}
\description{
Un-normalized log posterior of random effects theta_i
}
