% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{posterior_theta_nlme}
\alias{posterior_theta_nlme}
\title{Un-normalized log posterior of random effects theta_i}
\usage{
posterior_theta_nlme(
  xx,
  xx_names,
  par_list,
  long_loglike,
  model1,
  sub_dat1,
  Sigma,
  loc,
  theta_mu
)
}
\description{
Un-normalized log posterior of random effects theta_i
}
