% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_jm}
\alias{predict_jm}
\title{Predict longitudinal data based on JMStEM output}
\usage{
predict_jm(
  jm_res,
  kk,
  id = "PATIENT",
  par_est = NULL,
  time = "time",
  smooth = F,
  method = "mean"
)
}
\description{
Predict longitudinal data based on JMStEM output
}
