% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_saem}
\alias{predict_saem}
\title{Predict longitudinal data based on Monolix output}
\usage{
predict_saem(theta_dat, jm_res, id = "PATIENT", time = "week", smooth = F)
}
\description{
Predict longitudinal data based on Monolix output
}
