% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_nlme_pars.R
\name{score_fixed_par}
\alias{score_fixed_par}
\title{Gradient of negative log likelihood}
\usage{
score_fixed_par(
  xx,
  xx_names1,
  X2i_list,
  model1,
  loglike,
  randeff,
  par_list,
  new_dat,
  id,
  uniqueID,
  loc1
)
}
\description{
Gradient of negative log likelihood
}
