% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate longitudinal and survival data given specified model frameworks}
\usage{
simulate_data(
  n = 300,
  truncT = 24,
  model1,
  model2,
  covariance.model,
  cumhaz0 = c(rep(0.00015, 70), rep(5e-04, 70), rep(0.00015, 24 * 7 + 1 - 140)),
  data_frequency = NULL,
  age_par = c(41, 8),
  sex_par = 0.93,
  infection_type_par = 0.3,
  nadir_cd4_par = 0.3,
  nnrti_based_par = 0.5,
  add_var = T
)
}
\arguments{
\item{n}{sample size}

\item{truncT}{time when longitudinal data are truncated (in weeks)}

\item{model1}{NLME model for longitudinal data}

\item{model2}{Cox PH model for survival data}

\item{covariance.model}{a list specifying variance-covariance matrix structure for random parameters}

\item{cumhaz0}{cumulative hazard function for simulation survival data}

\item{data_frequency}{a vector specifying how frequent (in days) longitudinal data are collected}

\item{add_var}{whether adding variability in data_frequency or not}
}
\description{
Simulate longitudinal and survival data given specified model frameworks
}
