% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{trim_data}
\alias{trim_data}
\title{Trim longitudinal data based on desired data_frequency}
\usage{
trim_data(data, data_frequency, n, add_var = F)
}
\description{
Trim longitudinal data based on desired data_frequency
}
