% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_parameters.R
\name{update_parameters}
\alias{update_parameters}
\title{Update parameter estimates}
\usage{
update_parameters(
  sampling_output,
  cox_fm,
  missing_vars,
  loc,
  model1,
  model2,
  Sigma,
  S,
  n,
  id,
  uniqueID,
  RespLog,
  fixed_names1,
  random_effects,
  covariance.model,
  X1i_list,
  X2i_list,
  all_cases,
  par_list,
  mc.cores,
  sigma_par
)
}
\description{
Update parameter estimates
}
